Alteração retorno método getAltitude; Comentado código eixo;

+++ JavaApp/src/br/edu/ifba/embarcados/javaapp/conector/IComunicacaoSensores.java
@@ -7,7 +7,7 @@ public interface IComunicacaoSensores extends Library{
 	public int iniciar (String porta);
 	public int ler();
 	public char getQueda();
-	public float getAltitude();
+	public int getAltitude();
 	public int finalizar();
